requirements.txt
fastapi
uvicorn
jinja2
python-dotenv
httpx
mysql-connector-python
pydantic
python-multipart